Wrap the core styling into a mixin for reusability
It seems better to me to have a mixin for the core styles so that it becomes possible to easily turn other elements into icons (notably pseudo-elements), without having to locally duplicate the styles. This minor change makes that possible by letting people use `@include font-awesome-icon();`.

There is no change to the compiled CSS code. This code works on SassC and Ruby Sass.